Semana6/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==5.0.0